Move methods from MXSession to LoginRestClient
